moving npm based formatters from lib-extra to lib
and adapt dependencies where needed.
