+++ lib-extra/build.gradle
@@ -23,7 +23,3 @@ dependencies {
 
 // we'll hold the core lib to a high standard
 spotbugs { reportLevel = 'low' } // low|medium|high (low = sensitive to even minor mistakes)
-
-test { useJUnit { excludeCategories 'com.diffplug.spotless.category.NpmTest' } }
-
-task npmTest(type: Test) { useJUnit { includeCategories 'com.diffplug.spotless.category.NpmTest' } }
