+++ lib/src/main/java/com/diffplug/spotless/npm/PrettierFormatterStep.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.npm;
+package com.diffplug.spotless.npm;
 
 import static java.util.Arrays.asList;
 import static java.util.Objects.requireNonNull;
@@ -53,7 +53,7 @@ public class PrettierFormatterStep {
 			super(stepName,
 					provisioner,
 					new NpmConfig(
-							readFileFromClasspath(PrettierFormatterStep.class, "/com/diffplug/spotless/extra/npm/prettier-package.json"),
+							readFileFromClasspath(PrettierFormatterStep.class, "/com/diffplug/spotless/npm/prettier-package.json"),
 							"prettier"),
 					buildDir,
 					npm);
