+++ lib/src/main/resources/com/diffplug/spotless/npm/prettier-package.json
