+++ lib/src/main/resources/com/diffplug/spotless/npm/tsfmt-package.json
